Fixed circular imports by moving enums to separate file. Removed unecessary imports
