+++ assignments/aer1216_fall2020_hw2_sim.py
@@ -21,7 +21,7 @@ import pybullet as p
 from gym_pybullet_drones.envs.CtrlAviary import CtrlAviary
 from gym_pybullet_drones.utils.Logger import Logger
 from gym_pybullet_drones.utils.utils import sync
-from gym_pybullet_drones.envs.BaseAviary import DroneModel
+from gym_pybullet_drones.utils.enums import DroneModel
 from aer1216_fall2020_hw2_ctrl import HW2Control
 
 DURATION = 10
