+++ gym_pybullet_drones/control/DSLPIDControl.py
@@ -4,7 +4,7 @@ import pybullet as p
 from scipy.spatial.transform import Rotation
 
 from gym_pybullet_drones.control.BaseControl import BaseControl
-from gym_pybullet_drones.envs.BaseAviary import DroneModel, BaseAviary
+from gym_pybullet_drones.utils.enums import DroneModel
 
 class DSLPIDControl(BaseControl):
     """PID control class for Crazyflies.
