+++ gym_pybullet_drones/envs/DynAviary.py
@@ -1,7 +1,8 @@
 import numpy as np
 from gym import spaces
 
-from gym_pybullet_drones.envs.BaseAviary import DroneModel, Physics, BaseAviary
+from gym_pybullet_drones.envs.BaseAviary import BaseAviary
+from gym_pybullet_drones.utils.enums import DroneModel, Physics
 from gym_pybullet_drones.utils.utils import nnlsRPM
 
 class DynAviary(BaseAviary):
