+++ gym_pybullet_drones/envs/VisionAviary.py
@@ -2,7 +2,8 @@ import os
 import numpy as np
 from gym import spaces
 
-from gym_pybullet_drones.envs.BaseAviary import DroneModel, Physics, ImageType, BaseAviary
+from gym_pybullet_drones.envs.BaseAviary import BaseAviary
+from gym_pybullet_drones.utils.enums import DroneModel, Physics, ImageType
 
 class VisionAviary(BaseAviary):
     """Multi-drone environment class for control applications using vision."""
