+++ gym_pybullet_drones/envs/multi_agent_rl/FlockAviary.py
@@ -1,11 +1,7 @@
 import math
 import numpy as np
-from gym import spaces
-from ray.rllib.env.multi_agent_env import MultiAgentEnv, ENV_STATE
 
-from gym_pybullet_drones.envs.BaseAviary import DroneModel, Physics, BaseAviary
-
-from gym_pybullet_drones.envs.BaseAviary import DroneModel, Physics, BaseAviary
+from gym_pybullet_drones.utils.enums import DroneModel, Physics
 from gym_pybullet_drones.envs.single_agent_rl.BaseSingleAgentAviary import ActionType, ObservationType
 from gym_pybullet_drones.envs.multi_agent_rl.BaseMultiagentAviary import BaseMultiagentAviary
