+++ gym_pybullet_drones/envs/single_agent_rl/BaseSingleAgentAviary.py
@@ -1,12 +1,11 @@
 import os
-from datetime import datetime
 from enum import Enum
 import numpy as np
 from gym import spaces
 import pybullet as p
-import pybullet_data
 
-from gym_pybullet_drones.envs.BaseAviary import DroneModel, Physics, ImageType, BaseAviary
+from gym_pybullet_drones.envs.BaseAviary import BaseAviary
+from gym_pybullet_drones.utils.enums import DroneModel, Physics, ImageType
 from gym_pybullet_drones.utils.utils import nnlsRPM
 from gym_pybullet_drones.control.DSLPIDControl import DSLPIDControl
 from gym_pybullet_drones.control.SimplePIDControl import SimplePIDControl
