+++ gym_pybullet_drones/envs/single_agent_rl/FlyThruGateAviary.py
@@ -1,9 +1,8 @@
 import os
 import numpy as np
-from gym import spaces
 import pybullet as p
 
-from gym_pybullet_drones.envs.BaseAviary import DroneModel, Physics, BaseAviary
+from gym_pybullet_drones.utils.enums import DroneModel, Physics
 from gym_pybullet_drones.envs.single_agent_rl.BaseSingleAgentAviary import ActionType, ObservationType, BaseSingleAgentAviary
