+++ gym_pybullet_drones/envs/single_agent_rl/TakeoffAviary.py
@@ -1,7 +1,6 @@
 import numpy as np
-from gym import spaces
 
-from gym_pybullet_drones.envs.BaseAviary import DroneModel, Physics, BaseAviary
+from gym_pybullet_drones.utils.enums import DroneModel, Physics
 from gym_pybullet_drones.envs.single_agent_rl.BaseSingleAgentAviary import ActionType, ObservationType, BaseSingleAgentAviary
 
 class TakeoffAviary(BaseSingleAgentAviary):
